requirements.txt
# 深度学习框架
torch==2.9.0
torchvision==0.19.0
torchaudio==2.9.0

# CLIP模型
git+https://github.com/openai/CLIP.git

# YOLOv8
ultralytics==8.0.0

# 向量检索
faiss-gpu==1.7.4

# 数据库
pymysql==1.1.0
sqlalchemy==2.0.0

# Web框架
flask==3.0.0
flask-cors==4.0.0

# 视频处理
opencv-python==4.8.0
pillow==10.0.0

# 工具库
numpy==1.24.0
pandas==2.0.0
tqdm==4.66.0